+++ src/main/java/spot/spot/domain/pay/controller/PayController.java
@@ -9,8 +9,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import spot.spot.domain.job.entity.Job;
-import spot.spot.domain.job.service.ClientService;
+import spot.spot.domain.job.command.entity.Job;
+import spot.spot.domain.job.command.service.ClientCommandService;
+import spot.spot.domain.job.query.service.ClientQueryService;
 import spot.spot.domain.pay.entity.dto.request.PayApproveRequestDto;
 import spot.spot.domain.pay.entity.dto.request.PayReadyRequestDto;
 import spot.spot.domain.pay.entity.dto.response.PayApproveResponseDto;
@@ -24,11 +25,12 @@ import spot.spot.domain.pay.service.PayService;
 public class PayController {
 
     private final PayService payService;
-    private final ClientService clientService;
+    private final ClientQueryService clientQueryService;
+    private final ClientCommandService clientCommandService;
 
     @PostMapping("/deposit")
     public PayApproveResponseDto payApprove(@Valid @RequestBody PayApproveRequestDto request, Authentication auth) {
-        Job job = clientService.findByTid(request.tid());
+        Job job = clientQueryService.findByTid(request.tid());
         return payService.payApprove(
                 auth.getName(),
                 job,
@@ -39,10 +41,10 @@ public class PayController {
     @Transactional
     @PostMapping("/ready")
     public PayReadyResponseDto payReady(@Valid @RequestBody PayReadyRequestDto request, Authentication auth) {
-        Job findJob = clientService.findById(request.jobId());
+        Job findJob = clientQueryService.findById(request.jobId());
         PayReadyResponseDto payReadyResponseDto = payService.payReady(auth.getName(), request.content(), request.amount(), request.point(), findJob);
         String tid = payReadyResponseDto.tid();
-        clientService.updateTidToJob(findJob, tid);
+        clientCommandService.updateTidToJob(findJob, tid);
         return payReadyResponseDto;
     }
 }
