+++ src/main/java/spot/spot/domain/review/service/ReviewService.java
@@ -3,8 +3,8 @@ package spot.spot.domain.review.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import spot.spot.domain.job.entity.Job;
-import spot.spot.domain.job.repository.jpa.JobRepository;
+import spot.spot.domain.job.command.entity.Job;
+import spot.spot.domain.job.query.repository.jpa.JobRepository;
 import spot.spot.domain.review.dto.request.ReviewRequestDto;
 import spot.spot.domain.review.dto.response.ReviewResponseDto;
 import spot.spot.domain.review.entity.Review;
