+++ src/test/java/spot/spot/domain/pay/controller/PayControllerExceptionTest.java
@@ -13,8 +13,8 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
-import spot.spot.domain.job.entity.Job;
-import spot.spot.domain.job.service.ClientService;
+import spot.spot.domain.job.command.entity.Job;
+import spot.spot.domain.job.command.service.ClientCommandService;
 import spot.spot.domain.pay.entity.PayHistory;
 import spot.spot.domain.pay.entity.PayStatus;
 import spot.spot.domain.pay.entity.dto.request.PayApproveRequestDto;
@@ -47,7 +47,7 @@ public class PayControllerExceptionTest {
     PayService payService;
 
     @MockitoBean
-    ClientService clientService;
+    ClientCommandService clientCommandService;
 
     @MockitoBean
     private ClientRegistrationRepository clientRegistrationRepository;
@@ -144,7 +144,7 @@ public class PayControllerExceptionTest {
         PayHistory payHistory = PayHistory.builder().payAmount(1000).payPoint(1000).worker("worker").payStatus(PayStatus.PENDING).build();
         PayReadyResponseDto res = PayReadyResponseDto.create("redirect_pc_url", "redirect_mobile_url", "T123131", payHistory);
         when(payService.payReady(anyString(), anyString(), anyInt(), anyInt(), any())).thenReturn(res);
-        when(clientService.updateTidToJob(any(), any())).thenReturn(new Job());
+        when(clientCommandService.updateTidToJob(any(), any())).thenReturn(new Job());
 
         ///when ///then
         mockMvc.perform(
@@ -166,7 +166,7 @@ public class PayControllerExceptionTest {
         PayHistory payHistory = PayHistory.builder().payAmount(1000).payPoint(1000).worker("worker").payStatus(PayStatus.PENDING).build();
         PayReadyResponseDto res = PayReadyResponseDto.create("redirect_pc_url", "redirect_mobile_url", "T123131", payHistory);
         when(payService.payReady(anyString(), anyString(), anyInt(), anyInt(), any())).thenReturn(res);
-        when(clientService.updateTidToJob(any(), any())).thenReturn(new Job());
+        when(clientCommandService.updateTidToJob(any(), any())).thenReturn(new Job());
 
         ///when ///then
         mockMvc.perform(
@@ -189,7 +189,7 @@ public class PayControllerExceptionTest {
         PayHistory payHistory = PayHistory.builder().payAmount(1000).payPoint(1000).worker("worker").payStatus(PayStatus.PENDING).build();
         PayReadyResponseDto res = PayReadyResponseDto.create("redirect_pc_url", "redirect_mobile_url", "T123131", payHistory);
         when(payService.payReady(anyString(), anyString(), anyInt(), anyInt(), any())).thenReturn(res);
-        when(clientService.updateTidToJob(any(), any())).thenReturn(new Job());
+        when(clientCommandService.updateTidToJob(any(), any())).thenReturn(new Job());
 
         ///when ///then
         mockMvc.perform(
@@ -212,7 +212,7 @@ public class PayControllerExceptionTest {
         PayHistory payHistory = PayHistory.builder().payAmount(1000).payPoint(1000).worker("worker").payStatus(PayStatus.PENDING).build();
         PayReadyResponseDto res = PayReadyResponseDto.create("redirect_pc_url", "redirect_mobile_url", "T123131", payHistory);
         when(payService.payReady(anyString(), anyString(), anyInt(), anyInt(), any())).thenReturn(res);
-        when(clientService.updateTidToJob(any(), any())).thenReturn(new Job());
+        when(clientCommandService.updateTidToJob(any(), any())).thenReturn(new Job());
 
         ///when ///then
         mockMvc.perform(
